+++ src/CQChartsBarChartPlot.cpp
@@ -730,7 +730,7 @@ fillColor() const
   CQChartsBarChartPlot *plot = qobject_cast<CQChartsBarChartPlot *>(plot_);
 
   if (plot->isSetHidden(i_))
-    c = CQUtil::blendColors(c, Qt::white, 0.5);
+    c = CQUtil::blendColors(c, key_->bgColor(), 0.5);
 
   return c;
 }